% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlr_scr.R
\name{dlr_scr_vec}
\alias{dlr_scr_vec}
\title{Tancredi-Sellers-Ulam Scorer (Vectorized)}
\usage{
dlr_scr_vec(responses, target = "WORLD")
}
\arguments{
\item{responses}{Character vector of participant responses.}

\item{target}{See \code{\link[=dlr_scr]{dlr_scr()}}.}
}
\value{
Numeric vector of the same length as \code{responses}.
}
\description{
Vectorized convenience wrapper around \code{\link[=dlr_scr]{dlr_scr()}}.
}
\examples{
dlr_scr_vec(c('DLROW', 'DLORW', 'XYZ', NA))

}
