% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmse_calc.R
\name{levenshtein_distance}
\alias{levenshtein_distance}
\title{Calculate Levenshtein (Edit) Distance}
\usage{
levenshtein_distance(s1, s2)
}
\arguments{
\item{s1}{Character string (correct word, e.g., "DLROW")}

\item{s2}{Character string (input word from participant)}
}
\value{
Integer edit distance
}
\description{
Computes the Levenshtein distance between two strings using dynamic
programming.
}
\examples{
levenshtein_distance("DLROW", "DLROW")
levenshtein_distance("DLROW", "DRLOW")
levenshtein_distance("DLROW", "WORLD")

}
