% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmse_calc.R
\docType{data}
\name{mmse_calc}
\alias{mmse_calc}
\alias{.MISSING}
\title{MMSE WORLD Backwards Scoring Functions}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.MISSING
}
\description{
R port of the ADCS EDC mmsecalc2.pm Perl module.
Original creation date: 2014-06-24
}
\keyword{datasets}
