% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmse_calc.R
\name{score_mundo_backwards}
\alias{score_mundo_backwards}
\title{Score MUNDO Backwards Response (Spanish)}
\usage{
score_mundo_backwards(response)
}
\arguments{
\item{response}{Character string of participant's response}
}
\value{
Integer score from 0 to 5
}
\description{
Convenience wrapper for scoring Spanish MUNDO backwards responses.
}
\examples{
score_mundo_backwards("ODNUM")
score_mundo_backwards("ODMUN")

}
