% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmse_calc.R
\name{score_world_backwards}
\alias{score_world_backwards}
\title{Score WORLD Backwards Response}
\usage{
score_world_backwards(response)
}
\arguments{
\item{response}{Character string of participant's response}
}
\value{
Integer score from 0 to 5
}
\description{
Convenience wrapper for scoring English WORLD backwards responses.
}
\examples{
score_world_backwards("DLROW")
score_world_backwards("DRLOW")
score_world_backwards("DLORW")

}
