% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzworld-package.R
\docType{package}
\name{zzworld-package}
\alias{zzworld}
\alias{zzworld-package}
\title{zzworld: WORLD Test Scoring and Edit Distance Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions for analyzing WORLD test responses in cognitive assessments. Computes edit distance between patient responses and the target word "WORLD", supports multiple scoring methods, and integrates with MMSE (Mini-Mental State Examination) calculations. Useful for clinical research in cognitive assessment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rgt47/zzworld}
  \item Report bugs at \url{https://github.com/rgt47/zzworld/issues}
}

}
\author{
\strong{Maintainer}: Ronald G. Thomas \email{rgthomas@ucsd.edu} (\href{https://orcid.org/0000-0003-1686-4965}{ORCID})

Authors:
\itemize{
  \item Ronald G. Thomas \email{rgthomas@ucsd.edu} (\href{https://orcid.org/0000-0003-1686-4965}{ORCID})
}

}
\keyword{internal}
